requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
requests==2.31.0
dnspython==2.4.2
groq==0.4.1
flask-socketio==5.3.5
python-socketio==5.10.0
reportlab==4.0.7
Pillow==10.1.0
stripe==7.8.0
